+++ Class/Request/request.go
@@ -4,6 +4,7 @@ import (
 	"api_tester/Entity"
 	"fmt"
 	"net/http"
+	URL "net/url"
 	"strings"
 
 	Console "api_tester/Class/Console"
@@ -12,6 +13,7 @@ import (
 // Options Type
 type Options struct {
 	HeaderParam []string
+	Parameter   []string
 }
 
 // Send
@@ -23,13 +25,19 @@ func Send(url string, method string, HttpOptions Options) *http.Response {
 
 	switch method {
 	case "GET":
-		return GetRequest(url, Options{HeaderParam: HttpOptions.HeaderParam})
+		return GetRequest(url, HttpOptions)
 	default:
-		return GetRequest(url, Options{HeaderParam: HttpOptions.HeaderParam})
+		return GetRequest(url, HttpOptions)
 	}
 }
 
 func GetRequest(url string, options Options) *http.Response {
+	// Url Parameters
+	if len(options.Parameter) > 0 {
+		params := ParametersBuilder(options.Parameter)
+		url = url + "?" + params.Encode()
+	}
+
 	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		Console.Error("Error creating Request")
@@ -63,6 +71,9 @@ func BuildOption(config Entity.Test) Options {
 	if len(config.Headers) > 0 {
 		options.HeaderParam = config.Headers
 	}
+	if len(config.Parameters) > 0 {
+		options.Parameter = config.Parameters
+	}
 
 	return options
 }
@@ -82,4 +93,21 @@ func HeadersBuilder(headers []string) http.Header {
 	}
 
 	return httpHeaders
+}
+
+// ParametersBuilder
+/*
+Adding URL Parameters from list of string provided by TestConfig Parameters field
+*/
+func ParametersBuilder(params []string) URL.Values {
+	urlParameters := URL.Values{}
+
+	if params != nil && len(params) > 0 {
+		for _, paramStr := range params {
+			keyValue := strings.Split(paramStr, "=")
+			urlParameters.Add(keyValue[0], keyValue[1])
+		}
+	}
+
+	return urlParameters
 }